r/(legacy)virtual_machine: refactoring the helpers into the _test file to make refactoring easier
