+++ azurerm/internal/services/compute/virtual_machine_managed_disks_resource_test.go
@@ -2,20 +2,15 @@ package compute_test
 
 import (
 	"fmt"
-	"net/http"
 	"os"
 	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-06-01/compute"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/terraform"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance/check"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 )
 
 // NOTE: Test `TestAccVirtualMachine_enableAnWithVM` requires a machine of size `D8_v3` which is large/expensive - you may wish to ignore this test"
@@ -442,32 +437,6 @@ func TestAccVirtualMachine_ultraSSD(t *testing.T) {
 			),
 		},
 	})
-}
-
-func testCheckAndStopVirtualMachine(vm *compute.VirtualMachine) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.VMClient
-		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
-
-		vmID, err := azure.ParseAzureResourceID(*vm.ID)
-		if err != nil {
-			return fmt.Errorf("Unable to parse virtual machine ID %s, %+v", *vm.ID, err)
-		}
-
-		name := vmID.Path["virtualMachines"]
-		resourceGroup := vmID.ResourceGroup
-
-		future, err := client.Deallocate(ctx, resourceGroup, name)
-		if err != nil {
-			return fmt.Errorf("Failed stopping virtual machine %q: %+v", resourceGroup, err)
-		}
-
-		if err = future.WaitForCompletionRef(ctx, client.Client); err != nil {
-			return fmt.Errorf("Failed long polling for the stop of virtual machine %q: %+v", resourceGroup, err)
-		}
-
-		return nil
-	}
 }
 
 func (VirtualMachineResource) basicLinuxMachine_managedDisk_withOsWriteAcceleratorEnabled(data acceptance.TestData, enabled string) string {
@@ -2061,85 +2030,6 @@ resource "azurerm_virtual_machine" "test" {
   }
 }
 `, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, data.RandomString, data.RandomString)
-}
-
-func testCheckVirtualMachineManagedDiskExists(managedDiskID *string, shouldExist bool) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		d, err := testGetVirtualMachineManagedDisk(managedDiskID)
-		if err != nil {
-			return fmt.Errorf("Error trying to retrieve Managed Disk %s, %+v", *managedDiskID, err)
-		}
-		if d.StatusCode == http.StatusNotFound && shouldExist {
-			return fmt.Errorf("Unable to find Managed Disk %s", *managedDiskID)
-		}
-		if d.StatusCode != http.StatusNotFound && !shouldExist {
-			return fmt.Errorf("Found unexpected Managed Disk %s", *managedDiskID)
-		}
-
-		return nil
-	}
-}
-
-func testLookupVirtualMachineManagedDiskID(vm *compute.VirtualMachine, diskName string, managedDiskID *string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		if osd := vm.StorageProfile.OsDisk; osd != nil {
-			if strings.EqualFold(*osd.Name, diskName) {
-				if osd.ManagedDisk != nil {
-					id, err := findVirtualMachineManagedDiskID(osd.ManagedDisk)
-					if err != nil {
-						return fmt.Errorf("Unable to parse Managed Disk ID for OS Disk %s, %+v", diskName, err)
-					}
-					*managedDiskID = id
-					return nil
-				}
-			}
-		}
-
-		for _, dataDisk := range *vm.StorageProfile.DataDisks {
-			if strings.EqualFold(*dataDisk.Name, diskName) {
-				if dataDisk.ManagedDisk != nil {
-					id, err := findVirtualMachineManagedDiskID(dataDisk.ManagedDisk)
-					if err != nil {
-						return fmt.Errorf("Unable to parse Managed Disk ID for Data Disk %s, %+v", diskName, err)
-					}
-					*managedDiskID = id
-					return nil
-				}
-			}
-		}
-
-		return fmt.Errorf("Unable to locate disk %s on vm %s", diskName, *vm.Name)
-	}
-}
-
-func findVirtualMachineManagedDiskID(md *compute.ManagedDiskParameters) (string, error) {
-	if _, err := azure.ParseAzureResourceID(*md.ID); err != nil {
-		return "", err
-	}
-	return *md.ID, nil
-}
-
-func testGetVirtualMachineManagedDisk(managedDiskID *string) (*compute.Disk, error) {
-	client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.DisksClient
-	ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
-
-	armID, err := azure.ParseAzureResourceID(*managedDiskID)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to parse Managed Disk ID %s, %+v", *managedDiskID, err)
-	}
-	name := armID.Path["disks"]
-	resourceGroup := armID.ResourceGroup
-
-	d, err := client.Get(ctx, resourceGroup, name)
-	// check status first since sdk client returns error if not 200
-	if d.Response.StatusCode == http.StatusNotFound {
-		return &d, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	return &d, nil
 }
 
 func (VirtualMachineResource) linuxNoConfig(data acceptance.TestData) string {
