+++ azurerm/internal/services/compute/virtual_machine_resource_test.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+	"strings"
 	"testing"
 
+	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-06-01/compute"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
@@ -14,6 +16,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance/check"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
+	"github.com/tombuildsstuff/giovanni/storage/2019-12-12/blob/blobs"
 )
 
 type VirtualMachineResource struct {
@@ -141,6 +144,201 @@ func (t VirtualMachineResource) Exists(ctx context.Context, clients *clients.Cli
 	}
 
 	return utils.Bool(resp.ID != nil), nil
+}
+
+func testCheckVirtualMachineManagedDiskExists(managedDiskID *string, shouldExist bool) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		d, err := testGetVirtualMachineManagedDisk(managedDiskID)
+		if err != nil {
+			return fmt.Errorf("Error trying to retrieve Managed Disk %s, %+v", *managedDiskID, err)
+		}
+		if d.StatusCode == http.StatusNotFound && shouldExist {
+			return fmt.Errorf("Unable to find Managed Disk %s", *managedDiskID)
+		}
+		if d.StatusCode != http.StatusNotFound && !shouldExist {
+			return fmt.Errorf("Found unexpected Managed Disk %s", *managedDiskID)
+		}
+
+		return nil
+	}
+}
+
+func testLookupVirtualMachineManagedDiskID(vm *compute.VirtualMachine, diskName string, managedDiskID *string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		if osd := vm.StorageProfile.OsDisk; osd != nil {
+			if strings.EqualFold(*osd.Name, diskName) {
+				if osd.ManagedDisk != nil {
+					id, err := findVirtualMachineManagedDiskID(osd.ManagedDisk)
+					if err != nil {
+						return fmt.Errorf("Unable to parse Managed Disk ID for OS Disk %s, %+v", diskName, err)
+					}
+					*managedDiskID = id
+					return nil
+				}
+			}
+		}
+
+		for _, dataDisk := range *vm.StorageProfile.DataDisks {
+			if strings.EqualFold(*dataDisk.Name, diskName) {
+				if dataDisk.ManagedDisk != nil {
+					id, err := findVirtualMachineManagedDiskID(dataDisk.ManagedDisk)
+					if err != nil {
+						return fmt.Errorf("Unable to parse Managed Disk ID for Data Disk %s, %+v", diskName, err)
+					}
+					*managedDiskID = id
+					return nil
+				}
+			}
+		}
+
+		return fmt.Errorf("Unable to locate disk %s on vm %s", diskName, *vm.Name)
+	}
+}
+
+func findVirtualMachineManagedDiskID(md *compute.ManagedDiskParameters) (string, error) {
+	if _, err := azure.ParseAzureResourceID(*md.ID); err != nil {
+		return "", err
+	}
+	return *md.ID, nil
+}
+
+func testGetVirtualMachineManagedDisk(managedDiskID *string) (*compute.Disk, error) {
+	client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.DisksClient
+	ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
+
+	armID, err := azure.ParseAzureResourceID(*managedDiskID)
+	if err != nil {
+		return nil, fmt.Errorf("Unable to parse Managed Disk ID %s, %+v", *managedDiskID, err)
+	}
+	name := armID.Path["disks"]
+	resourceGroup := armID.ResourceGroup
+
+	d, err := client.Get(ctx, resourceGroup, name)
+	// check status first since sdk client returns error if not 200
+	if d.Response.StatusCode == http.StatusNotFound {
+		return &d, nil
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	return &d, nil
+}
+
+func testCheckAndStopVirtualMachine(vm *compute.VirtualMachine) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.VMClient
+		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
+
+		vmID, err := azure.ParseAzureResourceID(*vm.ID)
+		if err != nil {
+			return fmt.Errorf("Unable to parse virtual machine ID %s, %+v", *vm.ID, err)
+		}
+
+		name := vmID.Path["virtualMachines"]
+		resourceGroup := vmID.ResourceGroup
+
+		future, err := client.Deallocate(ctx, resourceGroup, name)
+		if err != nil {
+			return fmt.Errorf("Failed stopping virtual machine %q: %+v", resourceGroup, err)
+		}
+
+		if err = future.WaitForCompletionRef(ctx, client.Client); err != nil {
+			return fmt.Errorf("Failed long polling for the stop of virtual machine %q: %+v", resourceGroup, err)
+		}
+
+		return nil
+	}
+}
+
+func testCheckVirtualMachineVHDExistence(blobName string, shouldExist bool) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		storageClient := acceptance.AzureProvider.Meta().(*clients.Client).Storage
+		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
+
+		for _, rs := range s.RootModule().Resources {
+			if rs.Type != "azurerm_storage_container" {
+				continue
+			}
+
+			accountName := rs.Primary.Attributes["storage_account_name"]
+			containerName := rs.Primary.Attributes["name"]
+
+			account, err := storageClient.FindAccount(ctx, accountName)
+			if err != nil {
+				return fmt.Errorf("Error retrieving Account %q for Blob %q (Container %q): %s", accountName, blobName, containerName, err)
+			}
+			if account == nil {
+				return fmt.Errorf("Unable to locate Storage Account %q!", accountName)
+			}
+
+			client, err := storageClient.BlobsClient(ctx, *account)
+			if err != nil {
+				return fmt.Errorf("Error building Blobs Client: %s", err)
+			}
+
+			input := blobs.GetPropertiesInput{}
+			props, err := client.GetProperties(ctx, accountName, containerName, blobName, input)
+			if err != nil {
+				if utils.ResponseWasNotFound(props.Response) {
+					if !shouldExist {
+						return nil
+					}
+
+					return fmt.Errorf("The Blob for the Unmanaged Disk %q should exist in the Container %q but it didn't!", blobName, containerName)
+				}
+
+				return fmt.Errorf("Error retrieving properties for Blob %q (Container %q): %s", blobName, containerName, err)
+			}
+
+			if !shouldExist {
+				return fmt.Errorf("The Blob for the Unmanaged Disk %q shouldn't exist in the Container %q but it did!", blobName, containerName)
+			}
+		}
+
+		return nil
+	}
+}
+
+func testCheckVirtualMachineDisappears(resourceName string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.VMClient
+		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
+
+		// Ensure we have enough information in state to look up in API
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return fmt.Errorf("Not found: %s", resourceName)
+		}
+
+		vmName := rs.Primary.Attributes["name"]
+		resourceGroup, hasResourceGroup := rs.Primary.Attributes["resource_group_name"]
+		if !hasResourceGroup {
+			return fmt.Errorf("Bad: no resource group found in state for virtual machine: %s", vmName)
+		}
+
+		// this is a preview feature we don't want to use right now
+		var forceDelete *bool = nil
+		future, err := client.Delete(ctx, resourceGroup, vmName, forceDelete)
+		if err != nil {
+			return fmt.Errorf("Bad: Delete on vmClient: %+v", err)
+		}
+
+		if err = future.WaitForCompletionRef(ctx, client.Client); err != nil {
+			return fmt.Errorf("Bad: Delete on vmClient: %+v", err)
+		}
+
+		return nil
+	}
+}
+
+func testAccCheckVirtualMachineRecreated(t *testing.T, before, after *compute.VirtualMachine) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		if before.ID == after.ID {
+			t.Fatalf("Expected change of Virtual Machine IDs, but both were %v", before.ID)
+		}
+		return nil
+	}
 }
 
 func (VirtualMachineResource) winTimeZone(data acceptance.TestData) string {
