+++ azurerm/internal/services/compute/virtual_machine_unmanaged_disks_resource_test.go
@@ -11,9 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance/check"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
-	"github.com/tombuildsstuff/giovanni/storage/2019-12-12/blob/blobs"
 )
 
 func TestAccVirtualMachine_basicLinuxMachine(t *testing.T) {
@@ -2943,94 +2940,4 @@ resource "azurerm_virtual_machine" "test" {
   }
 }
 `, data.RandomInteger, data.Locations.Primary, data.RandomInteger, data.RandomInteger, data.RandomInteger, data.RandomInteger, data.RandomInteger)
-}
-
-func testCheckVirtualMachineVHDExistence(blobName string, shouldExist bool) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		storageClient := acceptance.AzureProvider.Meta().(*clients.Client).Storage
-		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
-
-		for _, rs := range s.RootModule().Resources {
-			if rs.Type != "azurerm_storage_container" {
-				continue
-			}
-
-			accountName := rs.Primary.Attributes["storage_account_name"]
-			containerName := rs.Primary.Attributes["name"]
-
-			account, err := storageClient.FindAccount(ctx, accountName)
-			if err != nil {
-				return fmt.Errorf("Error retrieving Account %q for Blob %q (Container %q): %s", accountName, blobName, containerName, err)
-			}
-			if account == nil {
-				return fmt.Errorf("Unable to locate Storage Account %q!", accountName)
-			}
-
-			client, err := storageClient.BlobsClient(ctx, *account)
-			if err != nil {
-				return fmt.Errorf("Error building Blobs Client: %s", err)
-			}
-
-			input := blobs.GetPropertiesInput{}
-			props, err := client.GetProperties(ctx, accountName, containerName, blobName, input)
-			if err != nil {
-				if utils.ResponseWasNotFound(props.Response) {
-					if !shouldExist {
-						return nil
-					}
-
-					return fmt.Errorf("The Blob for the Unmanaged Disk %q should exist in the Container %q but it didn't!", blobName, containerName)
-				}
-
-				return fmt.Errorf("Error retrieving properties for Blob %q (Container %q): %s", blobName, containerName, err)
-			}
-
-			if !shouldExist {
-				return fmt.Errorf("The Blob for the Unmanaged Disk %q shouldn't exist in the Container %q but it did!", blobName, containerName)
-			}
-		}
-
-		return nil
-	}
-}
-
-func testCheckVirtualMachineDisappears(resourceName string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		client := acceptance.AzureProvider.Meta().(*clients.Client).Compute.VMClient
-		ctx := acceptance.AzureProvider.Meta().(*clients.Client).StopContext
-
-		// Ensure we have enough information in state to look up in API
-		rs, ok := s.RootModule().Resources[resourceName]
-		if !ok {
-			return fmt.Errorf("Not found: %s", resourceName)
-		}
-
-		vmName := rs.Primary.Attributes["name"]
-		resourceGroup, hasResourceGroup := rs.Primary.Attributes["resource_group_name"]
-		if !hasResourceGroup {
-			return fmt.Errorf("Bad: no resource group found in state for virtual machine: %s", vmName)
-		}
-
-		// this is a preview feature we don't want to use right now
-		var forceDelete *bool = nil
-		future, err := client.Delete(ctx, resourceGroup, vmName, forceDelete)
-		if err != nil {
-			return fmt.Errorf("Bad: Delete on vmClient: %+v", err)
-		}
-
-		if err = future.WaitForCompletionRef(ctx, client.Client); err != nil {
-			return fmt.Errorf("Bad: Delete on vmClient: %+v", err)
-		}
-
-		return nil
-	}
-}
-
-func testAccCheckVirtualMachineRecreated(t *testing.T, before, after *compute.VirtualMachine) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		if before.ID == after.ID {
-			t.Fatalf("Expected change of Virtual Machine IDs, but both were %v", before.ID)
-		}
-		return nil
-	}
 }
